+++ src/client/pythonEnvironments/base/info/interpreter.ts
@@ -6,6 +6,7 @@ import {
     interpreterInfo as getInterpreterInfoCommand,
     InterpreterInfoJson,
 } from '../../../common/process/internal/scripts';
+import { IDisposable } from '../../../common/types';
 import { Architecture } from '../../../common/utils/platform';
 import { copyPythonExecInfo, PythonExecInfo } from '../../exec';
 import { parseVersion } from './pythonVersion';
@@ -45,7 +46,7 @@ type ShellExecResult = {
     stdout: string;
     stderr?: string;
 };
-type ShellExecFunc = (command: string, timeout: number) => Promise<ShellExecResult>;
+type ShellExecFunc = (command: string, timeout: number, disposables?: Set<IDisposable>) => Promise<ShellExecResult>;
 
 type Logger = {
     info(msg: string): void;
@@ -64,6 +65,7 @@ export async function getInterpreterInfo(
     python: PythonExecInfo,
     shellExec: ShellExecFunc,
     logger?: Logger,
+    disposables?: Set<IDisposable>,
 ): Promise<InterpreterInformation | undefined> {
     const [args, parse] = getInterpreterInfoCommand();
     const info = copyPythonExecInfo(python, args);
@@ -78,7 +80,7 @@ export async function getInterpreterInfo(
     // See these two bugs:
     // https://github.com/microsoft/vscode-python/issues/7569
     // https://github.com/microsoft/vscode-python/issues/7760
-    const result = await shellExec(quoted, 15000);
+    const result = await shellExec(quoted, 15000, disposables);
     if (result.stderr) {
         if (logger) {
             logger.error(`Failed to parse interpreter information for ${argv} stderr: ${result.stderr}`);
